Assignment completed upto custom exceptions

+++ restassignmen/src/main/java/com/rest/assignment/restassignment/controllers/StudentController.java
@@ -1,6 +1,7 @@
 package com.rest.assignment.restassignment.controllers;
 
 import com.rest.assignment.restassignment.entity.Student;
+import com.rest.assignment.restassignment.exceptions.StudentNotFoundException;
 import com.rest.assignment.restassignment.services.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -9,7 +10,6 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 public class StudentController {
@@ -24,9 +24,13 @@ public class StudentController {
     }
     
     @GetMapping("/students/{id}")
-    public Optional<Student> getSingleStudent(@PathVariable int id) {
-        Optional<Student> student = studentService.getSingleStudent(id);
-        return student;
+    public Student getSingleStudent(@PathVariable int id) {
+        Student student = studentService.getSingleStudent(id);
+        
+        if(student == null) {
+            throw new StudentNotFoundException("Student with id "+id+" does not exist.");
+        }
+        return studentService.getSingleStudent(id);
     }
     
     @PostMapping("/students")
@@ -45,6 +49,11 @@ public class StudentController {
     
     @DeleteMapping("/students/{id}")
     public void deleteStudent(@PathVariable int id) {
+        
+        Student student = studentService.getSingleStudent(id);
+        if(student == null ) {
+            throw new StudentNotFoundException("Student to be deleted is not present.");
+        }
         studentService.deleteStudent(id);
     }
